+++ app/models/serializers/work_queue/correspondence_document_serializer.rb
@@ -6,11 +6,11 @@ class WorkQueue::CorrespondenceDocumentSerializer
   attribute :correspondence_id
   attribute :document_file_number
   attribute :pages
-  attribute :vbms_document_id
+  attribute :vbms_document_type_id
   attribute :uuid
   attribute :document_type
   attribute :document_title do |object|
-    doc_id = object.attributes["document_type"]
+    doc_id = object.attributes["vbms_document_type_id"]
     Caseflow::DocumentTypes::TYPES[doc_id]
   end
 end
